refactor(core) add kong.concurrency module
Introduces a new module, `kong.concurrency`, with high-level concurrency
constructs.

Kong has a layered concurrency structure (a cluster of nodes, where each node
has a pool of workers, where each worker has a set of coroutines). Some
operations need to be performed in a mutually-exclusive manner with regard to
one of these layers; code for implementing this in each level uses different
low-level constructs and can be tricky to get right (it has been a source of
bugs in the past).

This module unifies the interfaces for concurrency constructs in these levels.