feat(api) create immediate child entities with single admin api call
Makes it possible to execute admin api calls such as:
http :8001/routes paths=/ \
                  service.host=example.org \
                  service.path=/ \
                  service.port=80 -f

Passing service.url doesn't work atm, but becomes available when this
is merged with recent work that makes so called shorthands available
from schemas.

At its current form this is just a PoC. The biggest thing is to think
about the correct way to do error reporting.